Add files via upload
Changed synergyTheory,R so it has everything needed for the LSSR paper except the cross validation chunk, where I am having trouble sub-setting data-frames. Everything that works outside of a for loop fails inside the for loop except inconvenient subsetting by just using the actual integers. For example forcing a variable that used to be part of a data frame to be an integer and then trying to use it in the for loop instead of typing in "4" gave an error message.